CUDA optimizations.

+++ hash/gpu/cuda/cuda_hasher.h
@@ -17,9 +17,9 @@ struct cuda_kernel_arguments {
     int32_t *address_profile_1_1_524288;
     uint32_t *address_profile_4_4_16384;
     uint32_t *segments_profile_4_4_16384;
-    uint64_t *seed_memory[2];
-    uint64_t *out_memory[2];
-    uint64_t *host_seed_memory[2];
+    uint32_t *seed_memory[2];
+    uint32_t *out_memory[2];
+    uint32_t *host_seed_memory[2];
 };
 
 struct argon2profile_info {
